+++ internal/cloudformation/types.go
@@ -1,0 +1,8 @@
+package cloudformation
+
+type Violation struct {
+	ResourceName string
+	ResourceType string
+	Line         int
+	MissingTags  []string
+}
